CCapture Fixes Round 1
