+++ components/CCapture/download.js
@@ -1,3 +1,5 @@
+/* credit to James Pollack for original code modifications - https://github.com/imgntn/j360 */
+
 //download.js v3.0, by dandavis; 2008-2014. [CCBY2] see http://danml.com/download.html for tests/usage
 // v1 landed a FF+Chrome compat way of downloading strings to local un-named files, upgraded to use a hidden frame and optional mime
 // v2 added named files via a[download], msSaveBlob, IE (10+) support, and window.URL support for larger+faster saves than dataURLs
