+++ components/CCapture/gif.js
