+++ components/CCapture/gif.worker.js
