+++ index.html
@@ -157,8 +157,10 @@
         </script>
 
         <!-- The following statement is needed because the CCapture library seems to have not be written as an ES6-style module, so it can't be imported with an import statement -->
-        <script type="text/javascript" src="./CCapture.all.min.js"></script>
-        <!-- <script type="text/javascript" src="./ccapture_workers/CCapture.js"></script> -->
+        <script type="text/javascript" src="./components/CCapture/CCapture.js"></script>
+        <script type="text/javascript" src="./components/CCapture/download.js"></script>
+        <script type="text/javascript" src="./components/CCapture/tar.js"></script>
+        <script type="text/javascript" src="./components/CCapture/Whammy.js"></script>
 
         <script type="module" src="./main.js"></script>
         <div id="title">
@@ -174,10 +176,9 @@
             <input type="radio" name="framerate" id="encode-120" value="120" > <label for="encode-120" >120FPS</label></p>
           <p><b>Format</b><br/>
             <input type="radio" name="encoder" id="encode-webm" value="webm" > <label for="encode-webm" >WebM</label>
-            <input type="radio" name="encoder" id="encode-gif" value="gif" > <label for="encode-gif" >GIF</label>
+            <!-- <input type="radio" name="encoder" id="encode-gif" value="gif" > <label for="encode-gif" >GIF</label> -->
             <input type="radio" name="encoder" id="encode-png" value="png" checked="checked" > <label for="encode-png" >PNG</label>
             <input type="radio" name="encoder" id="encode-jpg" value="jpg" > <label for="encode-jpg" >JPEG</label></p>
-            <input type="radio" name="encoder" id="encode-webm-mediarecorder" value="webm-mediarecorder" > <label for="encode-webm-mediarecorder" >WebM (MediaRecorder) <b>WIP</b></label>
           <p><b>Other options</b><br/>
             <input id="apply-motionblur" name="motion-blur" type="checkbox" /> <label for="apply-motionblur">Apply motion blur</label>
           </p>
@@ -195,10 +196,6 @@
 
     </div>
 
-  
-    <!-- <script type="text/javascript" src="../../build/CCapture.all.min.js"></script> -->
-    <!-- <script type="module" src="./main.js"></script> -->
-  
   </body>
 </html>
